refactor(requestcontext): align package name and godoc comments
